frontApp/backapp/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3

# Django core
Django>=4.2.0,<5.0.0

# REST Framework and extensions
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
djoser>=2.1.0
djangorestframework-simplejwt>=5.2.0

# Real-time features
channels>=4.0.0
channels-redis>=4.1.0

# Database adapters
psycopg2-binary>=2.9.5  # For PostgreSQL
# mysqlclient>=2.1.1  # For MySQL (uncomment if needed)

# API Documentation
drf-yasg>=1.21.4

# Cross-origin resource sharing
django-cors-headers>=3.13.0

# Filtering
django-filter>=22.1

# Forms and UI
django-crispy-forms>=1.14.0

# Authentication and user management
django-allauth>=0.51.0

# Development tools
ipython>=8.5.0
django-debug-toolbar>=3.7.0

# Date/time handling
pytz>=2022.5

# Image processing
Pillow>=9.2.0

# Utilities
python-dotenv>=0.21.0